added useAllUsers hook
